app/requirements.txt
streamlit
numpy
pandas
scikit-learn
joblib
sqlalchemy
psycopg2-binary
python-dotenv
reportlab
matplotlib
